pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "msl-tools"
version = "0.1.0"
description = "Tools for working with Markdown Specification Language (MSL)"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "MSL Contributors"},
]
maintainers = [
    {name = "MSL Contributors"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Documentation",
    "Topic :: Text Processing :: Markup :: Markdown",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["markdown", "specification", "documentation", "requirements"]
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
render = [
    "Jinja2>=3.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/chrs-myrs/msl-specification"
Documentation = "https://github.com/chrs-myrs/msl-specification/tree/master/docs"
Issues = "https://github.com/chrs-myrs/msl-specification/issues"
Repository = "https://github.com/chrs-myrs/msl-specification"

[project.scripts]
msl-lint = "tools.cli.msl_lint:main"
msl-render = "tools.cli.msl_render:main"

[tool.setuptools]
packages = ["tools", "tools.lib", "tools.cli"]

[tool.setuptools.package-data]
msl = ["schemas/*.json", "schemas/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tools/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true